requirements.txt
asgiref==3.11.0
beautifulsoup4==4.14.3
bs4==0.0.2
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
cloudinary==1.44.1
cryptography==46.0.4
dj-database-url==3.1.0
Django==6.0
django-cloudinary-storage==0.3.0
django-cors-headers==4.9.0
django-filter==25.2
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
idna==3.11
Markdown==3.10
MarkupSafe==3.0.3
pillow==12.1.0
psycopg2-binary==2.9.11
pycparser==3.0
PyJWT==2.10.1
python-dotenv==1.2.1
python-http-client==3.3.7
requests==2.32.5
sendgrid==6.12.5
six==1.17.0
soupsieve==2.8.3
sqlparse==0.5.5
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
Werkzeug==3.1.5
whitenoise==6.11.0
